% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.list.r
\docType{package}
\name{broom}
\alias{broom}
\alias{_PACKAGE}
\alias{broom.list}
\alias{broom.list-package}
\title{Convert Statistical List Output into Tidy Tibbles}
\description{
This package provides methods for the generic functions \code{tidy()}
and \code{glance()} for lists, which are often returned by statistical analysis
functions without S3 class attributes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/corybrunson/broom.list}
  \item Report bugs at \url{https://github.com/corybrunson/broom.list/issues}
}

}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID}) [contributor]

Authors:
\itemize{
  \item Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})
  \item Simon Couch \email{simonpatrickcouch@gmail.com} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Brian Fannin \email{captain@pirategrunt.com} [contributor]
  \item Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID}) [contributor]
  \item Nina Jakobsen [contributor]
}

}
\keyword{internal}
