% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-cmdscale-tidiers.r
\name{glance_cmdscale}
\alias{glance_cmdscale}
\alias{glance.cmdscale}
\title{Tidy a(n) cmdscale object masquerading as list}
\usage{
glance_cmdscale(x, ...)
}
\arguments{
\item{x}{A list with components \code{points}, \code{eig}, \code{x}, \code{ac}, and \code{GOF} as
returned by \code{\link[stats:cmdscale]{stats::cmdscale()}} under certain conditions.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with exactly one row and the following columns:

\item{n}{Number of points.}
\item{k}{Dimension of data representation space.}
\item{ac}{Additive constant \eqn{c}.}
\item{GOF1,GOF2}{Goodness of fit as calculated from the eigenvalues.}
}
\description{
Broom tidies a number of lists that are effectively S3
objects without a class attribute. For example, \code{\link[stats:optim]{stats::optim()}},
\link[base:svd]{svd()} and \code{\link[interp:interp]{interp::interp()}} produce consistent output, but
because they do not have a class attribute, they cannot be handled by S3
dispatch.

These functions look at the elements of a list and determine if there is
an appropriate tidying method to apply to the list. Those tidiers are
implemented as functions of the form \verb{tidy_<function>()} or
\verb{glance_<function>()} and are not exported (but they are documented!).

If no appropriate tidying method is found, they throw an error.
}
\examples{
# 'dist' object (matrix of road distances) of large American cities
class(UScitiesD)
print(UScitiesD)

# use multidimensional scaling to infer artificial planar coordinates
UScitiesD \%>\%
  cmdscale(k = 3L, eig = TRUE, x.ret = TRUE) ->
  usa_mds

# glance at the model
glance(usa_mds)
# return the tidied coordinates
tidy(usa_mds)
# return the upper triangle of the doubly-centered distance matrix in tidy form
tidy(usa_mds, matrix = "x")
# return the eigenvalues for the principal coordinates, with summary statistics
tidy(usa_mds, matrix = "eig")
# reorient to conventional compass
tidy(usa_mds) \%>\%
  ggplot(aes(x = -PCo1, y = -PCo2)) +
  geom_text(aes(label = point), size = 3) +
  ggtitle("MDS biplot of distances between U.S. cities")
}
\seealso{
\code{\link[generics:glance]{generics::glance()}}, \code{\link[stats:cmdscale]{stats::cmdscale()}}

Other list tidiers: 
\code{\link{glance_optim}()},
\code{\link{list_tidiers}},
\code{\link{tidy_cancor}()},
\code{\link{tidy_cmdscale}()},
\code{\link{tidy_irlba}()},
\code{\link{tidy_optim}()},
\code{\link{tidy_svd}()},
\code{\link{tidy_xyz}()}
}
\concept{list tidiers}
