% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.list.r, R/utils.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{tidy}
\alias{glance}
\alias{\%>\%}
\title{Objects exported from other packages}
\seealso{
\code{\link[=tidy.list]{tidy.list()}}

\code{\link[=glance.list]{glance.list()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

