% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-cancor-tidiers.r
\name{tidy_cancor}
\alias{tidy_cancor}
\alias{tidy.cancor}
\title{Tidy a(n) cancor object masquerading as list}
\usage{
tidy_cancor(x, matrix = "cor", ...)
}
\arguments{
\item{x}{A list with components \code{cor}, \code{xcoef}, \code{ycoef}, \code{xcenter}, and
\code{ycenter} as returned by \code{\link[stats:cancor]{stats::cancor()}}.}

\item{matrix}{Character specifying which list element should be tidied,
matched to the following options (defaults to \code{"cor"}):
\itemize{
\item \code{"xcoef"}: returns information about the estimated coefficients in the
\code{x} variables.
\item \code{"ycoef"}: returns information about the estimated coefficients in the
\code{y} variables
\item \code{"cor"}: returns information about the canonical correlations.
}}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns depending on the element of the CCA
being tidied.

If \code{matrix} is \code{"xcoef"} (respectively, \code{"ycoef"}), each row in the tidied
output corresponds to the estimated coefficients of the \code{x} (respectively,
\code{y}) variables. The columns are:

\item{\code{CC}}{Integer indicating a canonical dimension.}
\item{\code{variable}}{Column name (or integer position) in the original data.}
\item{\code{value}}{The coordinate of this variable in the canonical dimension.}

If \code{matrix} is \code{"cor"}, each row contains information about one canonical
dimension. The columns are:

\item{\code{CC}}{An integer indicating the canonical dimension.}
\item{\code{cor}}{The canonical correlation in this dimension.}
\item{\code{percent}}{Fraction of variation along this canonical dimension
(a numeric value between 0 and 1).}
\item{\code{cumulative}}{Cumulative fraction of variation along canonical
dimensions up to this one
(a numeric value between 0 and 1).}
}
\description{
Broom tidies a number of lists that are effectively S3
objects without a class attribute. For example, \code{\link[stats:optim]{stats::optim()}},
\link[base:svd]{svd()} and \code{\link[interp:interp]{interp::interp()}} produce consistent output, but
because they do not have a class attribute, they cannot be handled by S3
dispatch.

These functions look at the elements of a list and determine if there is
an appropriate tidying method to apply to the list. Those tidiers are
implemented as functions of the form \verb{tidy_<function>()} or
\verb{glance_<function>()} and are not exported (but they are documented!).

If no appropriate tidying method is found, they throw an error.
}
\examples{
# savings data
class(LifeCycleSavings)
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]
# canonical correlation analysis
savings_cca <- cancor(pop, oec)

# return the tidied canonical coefficients for the left variables
tidy(savings_cca)
# return the tidied canonical coefficients for the right variables
tidy(savings_cca, matrix = "ycoef")
# return the canonical coefficients, with summary statistics
tidy(savings_cca, matrix = "cor")

# scree plot
ggplot(tidy(savings_cca, matrix = "cor"), aes(x = CC, y = percent)) +
  theme_bw() +
  geom_col() +
  labs(x = "Canonical dimension", y = "Percent of variance")
}
\seealso{
\code{\link[generics:tidy]{generics::tidy()}}, \code{\link[stats:cancor]{stats::cancor()}}

Other list tidiers: 
\code{\link{glance_cmdscale}()},
\code{\link{glance_optim}()},
\code{\link{list_tidiers}},
\code{\link{tidy_cmdscale}()},
\code{\link{tidy_irlba}()},
\code{\link{tidy_optim}()},
\code{\link{tidy_svd}()},
\code{\link{tidy_xyz}()}
}
\concept{list tidiers}
