% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-cmdscale-tidiers.r
\name{tidy_cmdscale}
\alias{tidy_cmdscale}
\alias{tidy.cmdscale}
\title{Tidy a(n) cmdscale object masquerading as list}
\usage{
tidy_cmdscale(x, matrix = "points", ...)
}
\arguments{
\item{x}{A list with components \code{points}, \code{eig}, \code{x}, \code{ac}, and \code{GOF} as
returned by \code{\link[stats:cmdscale]{stats::cmdscale()}} under certain conditions.}

\item{matrix}{Character specifying which list element should be tidied,
matched to the following options (defaults to \code{"points"}):
\itemize{
\item \code{"points"}: returns information about the coordinates in the
representation space.
\item \code{"x"}: returns information about the doubly-centered symmetric matrix
used in the calculation.
\item \code{"eig"}: returns information about the eigenvalues.
}}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns depending on the element of the CMDS
being tidied.

If \code{matrix} is \code{"points"}, each row corresponds to one object (point) in
the original data object. The columns are:

\item{\code{point}}{ID (character name or integer number) of the point in the
original data.}
\item{\verb{PCo*}}{The principal coordinates of this point.}

If \code{matrix} is \code{"x"}, each row corresponds to one pair of points in the
original data. The columns are:

\item{\code{point1}}{The first point name, in order of the original data.}
\item{\code{point2}}{The second point name, in order of the original data.}
\item{\code{DC}}{The corresponding entry in the doubly-centered symmetric
matrix.}

If \code{matrix} is \code{"eig"}, each row contains information about one dimension
of the eigendecomposition that powers the scaling process. The columns are:

\item{\code{PCo}}{An integer indicating the principal coordinate.}
\item{\code{eig}}{The eigenvalue along this principal coordinate.}
\item{\code{percent}}{Fraction of variation along this principal coordinate
(a numeric value between 0 and 1).}
\item{\code{cumulative}}{Cumulative fraction of variation along principal
coordinates up to this one
(a numeric value between 0 and 1).}
}
\description{
Broom tidies a number of lists that are effectively S3
objects without a class attribute. For example, \code{\link[stats:optim]{stats::optim()}},
\link[base:svd]{svd()} and \code{\link[interp:interp]{interp::interp()}} produce consistent output, but
because they do not have a class attribute, they cannot be handled by S3
dispatch.

These functions look at the elements of a list and determine if there is
an appropriate tidying method to apply to the list. Those tidiers are
implemented as functions of the form \verb{tidy_<function>()} or
\verb{glance_<function>()} and are not exported (but they are documented!).

If no appropriate tidying method is found, they throw an error.
}
\details{
When \code{\link[=cmdscale]{cmdscale()}} is instructed to return any of several optional elements,
or when \code{list. = TRUE}, the output is not the default point coordinate matrix
but a 5-element list with a consistent naming scheme (though some elements
will be \code{NULL} if their parameters are not set to \code{TRUE}). These tidiers rely
on this list structure to organize the model output into a tibble.
}
\examples{
# 'dist' object (matrix of road distances) of large American cities
class(UScitiesD)
print(UScitiesD)

# use multidimensional scaling to infer artificial planar coordinates
UScitiesD \%>\%
  cmdscale(k = 3L, eig = TRUE, x.ret = TRUE) ->
  usa_mds

# glance at the model
glance(usa_mds)
# return the tidied coordinates
tidy(usa_mds)
# return the upper triangle of the doubly-centered distance matrix in tidy form
tidy(usa_mds, matrix = "x")
# return the eigenvalues for the principal coordinates, with summary statistics
tidy(usa_mds, matrix = "eig")
# reorient to conventional compass
tidy(usa_mds) \%>\%
  ggplot(aes(x = -PCo1, y = -PCo2)) +
  geom_text(aes(label = point), size = 3) +
  ggtitle("MDS biplot of distances between U.S. cities")
}
\seealso{
\code{\link[generics:tidy]{generics::tidy()}}, \code{\link[stats:cmdscale]{stats::cmdscale()}}

Other list tidiers: 
\code{\link{glance_cmdscale}()},
\code{\link{glance_optim}()},
\code{\link{list_tidiers}},
\code{\link{tidy_cancor}()},
\code{\link{tidy_irlba}()},
\code{\link{tidy_optim}()},
\code{\link{tidy_svd}()},
\code{\link{tidy_xyz}()}
}
\concept{list tidiers}
